Finally got the model view projection as I wanted

+++ src/main/java/es/optocom/jovp/rendering/Processing.java
@@ -49,8 +49,8 @@ class Processing {
      *
      * @since 0.0.1
      */
-    public void frequency(float xp, float xf, float yp, float yf) {
-        this.frequency = new Vector4f(xp, xf, yp, yf);
+    public void frequency(double xp, double xf, double yp, double yf) {
+        this.frequency = new Vector4f((float) xp, (float) xf, (float) yp, (float) yf);
     }
 
     /**
@@ -64,8 +64,8 @@ class Processing {
      *
      * @since 0.0.1
      */
-    public void contrast(float r, float g, float b, float a) {
-        this.contrast = new Vector4f(r, g, b, a);
+    public void contrast(double r, double g, double b, double a) {
+        this.contrast = new Vector4f((float) r, (float) g, (float) b, (float) a);
     }
 
     /**
@@ -73,24 +73,25 @@ class Processing {
      * Rotate the texture inside the model
      *
      * @param rotation Angle of rotation in degrees
-     * @param pivot    Pivot UV values
+     * @param pivot Pivot UV values
      *
      * @since 0.0.1
      */
-    public void rotation(double rotation, float[] pivot) {
+    public void rotation(double rotation, double[] pivot) {
     }
 
     /**
      *
      * Add a Gaussian envelope
      *
-     * @param sdx   Defocus in diopters on the x-axis
-     * @param sdy   Defocus in diopters on the y-axis
-     * @param angle Angle
+     * @param type Envelope type
+     * @param sdx Envelope Gaussian standard deviation for x-axis in degrees
+     * @param sdy Envelope Gaussian standard deviation for y-axis in degrees
+     * @param angle Angle of rotation
      *
      * @since 0.0.1
      */
-    void envelope(EnvelopeType type, float sdx, float sdy, float angle) {
+    void envelope(EnvelopeType type, double sdx, double sdy, double angle) {
         settings.y = switch(type) {
             case NONE -> 0;
             case SQUARE -> 1;
@@ -106,17 +107,17 @@ class Processing {
      *
      * Add Gaussian defocus (spherical and astigmatic defocus)
      *
-     * @param dx    Defocus in diopters for the x-axis
-     * @param dy    Defocus in diopters for the y-axis
+     * @param dx Defocus in diopters for the x-axis
+     * @param dy  Defocus in diopters for the y-axis
      * @param angle Angle
      *
      * @since 0.0.1
      */
-    void defocus(float dx, float dy, float angle) {
+    void defocus(double dx, double dy, double angle) {
         settings.z = 1;
         // TODO: convert from Diopters to Gaussian SD in visual angle
-        float sdx = dx;
-        float sdy = dy;
+        double sdx = dx;
+        double sdy = dy;
         defocus.x = (float) Math.toRadians(sdx) / 2;
         defocus.y = (float) Math.toRadians(sdy) / 2;
         defocus.z = (float) Math.toRadians(angle);
